Add basic support to open your videos in process
This implementation add a custom variable `ivy-youtube-play-at` that
defines where the video will be played. The default value ís "browser"
to maintain compatibility with the last version.

You can set this variable to `/usr/bin/mpv` for example, if you do
that when you find a video it will be played on mpv instead browser.